Moved pipeline_messaging
