+++ epymorph/log/messaging.py
@@ -12,7 +12,7 @@ import humanize
 from humanize import naturalsize
 
 from epymorph.attribute import NAME_PLACEHOLDER
-from epymorph.event import ADRIOProgress, EventBus, OnStart, OnTick
+from epymorph.event import ADRIOProgress, EventBus, OnPipelineProgress, OnPipelineStart, OnStart, OnTick
 from epymorph.settings import declare_setting, env_flag
 from epymorph.util import progress, subscriptions
 
@@ -195,6 +195,119 @@ def sim_messaging(
         if live:
             subs.subscribe(_events.on_tick, on_tick)
         subs.subscribe(_events.on_finish, on_finish)
+        if adrio:
+            subs.subscribe(_events.on_adrio_progress, on_adrio_progress)
+        yield  # to outer context
+        # And now our event handlers will be unsubscribed.
+
+
+@contextmanager
+def pipeline_messaging(
+    *,
+    adrio: bool = True,
+    live: bool | None = None,
+) -> Generator[None, None, None]:
+    """
+    """
+
+    live = is_live_messaging(live)
+    start_time: float | None = None
+    total_progress: int = 0
+    progress_complete = 0
+
+    def on_pipeline_start(e: OnPipelineStart) -> None:
+        nonlocal total_progress
+        nonlocal progress_complete
+        progress_complete = 0
+        start_date = e.rume.time_frame.start_date
+        end_date = e.rume.time_frame.end_date
+        duration_days = e.rume.time_frame.days
+
+        total_progress = e.total_progress
+
+        print(f"Running simulation ({e.name}):")
+        print(f"• {start_date} to {end_date} ({duration_days} days)")
+        print(f"• {e.rume.scope.nodes} geo nodes")
+        print(f"• {e.num_realizations} realizations")
+        if live:
+            print(progress(0.0), end="\r")
+
+        nonlocal start_time
+        start_time = perf_counter()
+
+    # keeping track of the length of the last line we printed
+    # lets us clear any trailing characters when rendering stuff
+    # after the progress bar of varying width
+    last_progress_length = 0
+
+    def on_pipeline_progress(e: OnPipelineProgress) -> None:
+        nonlocal last_progress_length
+        nonlocal total_progress
+        nonlocal progress_complete
+        progress_complete += e.progress
+        total_process_time = perf_counter() - start_time
+        average_process_time = total_process_time / progress_complete
+
+        percent_complete = (progress_complete + 1) / total_progress
+        progress_left = total_progress - progress_complete
+
+        # multiply the remaining ticks by the average processing time
+        estimate = progress_left * average_process_time
+
+        time_remaining = humanize.precisedelta(ceil(estimate), minimum_unit="seconds")
+        formatted_time = f"({time_remaining} remaining)"
+        line = f"  {progress(percent_complete)}"
+        # if no time remaining, omit the time progress
+        if estimate > 0:
+            line += f"{formatted_time}"
+        print(line.ljust(last_progress_length), end="\r")
+        last_progress_length = len(line)
+
+    def on_pipeline_finish(_: None) -> None:
+        end_time = perf_counter()
+        line = f"  {progress(1.0)}"
+        print(line.ljust(last_progress_length), end="\n")
+        if start_time is not None:
+            print(f"Runtime: {(end_time - start_time):.3f}s")
+
+    def on_adrio_progress(e: ADRIOProgress) -> None:
+        nonlocal last_progress_length
+        if e.ratio_complete == 0:
+            if e.attribute == NAME_PLACEHOLDER:
+                print(f"Loading {e.adrio_name}:")
+            else:
+                print(f"Loading {e.attribute} ({e.adrio_name}):")
+        if not e.final:
+            if e.download is None:
+                dl = ""
+            else:
+                total, downloaded, speed = e.download
+                ff = partial(naturalsize, binary=False)  # format file size
+                dwn = "?" if downloaded is None else ff(downloaded)
+                tot = "?" if total is None else ff(total)
+                spd = "?" if speed is None else ff(speed)
+                dl = f" {dwn}/{tot} ({spd}/s)"
+
+            if live:
+                line = f"  {progress(e.ratio_complete)}{dl}"
+                print(line.ljust(last_progress_length), end="\r")
+                last_progress_length = len(line)
+        else:
+            if e.duration is None:
+                dur = ""
+            else:
+                dur = f" ({e.duration:0.3f}s)"
+            line = f"  {progress(e.ratio_complete)}{dur}"
+            print(line.ljust(last_progress_length))
+            last_progress_length = 0
+
+    with subscriptions() as subs:
+        # Set up a subscriptions context, subscribe our handlers,
+        # then yield to the outer context (ostensibly where the sim will be run).
+        subs.subscribe(_events.on_pipeline_start, on_pipeline_start)
+        if live:
+            subs.subscribe(_events.on_pipeline_progress, on_pipeline_progress)
+        subs.subscribe(_events.on_pipeline_finish, on_pipeline_finish)
         if adrio:
             subs.subscribe(_events.on_adrio_progress, on_adrio_progress)
         yield  # to outer context
